feat(ts): update MongooseLoader
