--- src/__tests__/index.spec.js
@@ -1,5 +1,0 @@
-/* @flow */
-// TODO - add real tests here
-it('should sum', () => {
-  expect(true).toBe(true);
-});
